add api of order

+++ api/common/src/main/java/com/zero/common/po/OrderDetail.java
@@ -1,16 +1,14 @@
 package com.zero.common.po;
 
 import io.swagger.annotations.ApiModelProperty;
-import lombok.Data;
-import lombok.ToString;
-
+import java.io.Serializable;
+import java.util.Date;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.io.Serializable;
-import java.math.BigDecimal;
-import java.util.Date;
+import lombok.Data;
+import lombok.ToString;
 
 @Data
 @ToString
@@ -28,7 +26,7 @@ public class OrderDetail implements Serializable {
     private String productName;
 
     @ApiModelProperty(value = "商品价格")
-    private BigDecimal productPrice;
+    private Double productPrice;
 
     @ApiModelProperty(value = "商品数量")
     private Integer productQuantity;
